sshd hardening + don't use root in
Added ansible-sshd role from github (check the requirements.yml for the
url).

Added configuration to disable root logins, logins with passwords, PAM.
Also restricted ssh logins to users in the admin groups which should
prevent logins with other 'special' users like nagios etc.

Improved the main "configure_system.yml" so that root is only used to
add ssh admin users, and only then if the current user account can't ssh
to the system.

We still need to remove the admin keys from the root
account, but this can be done later, and is not so critical now that
root logins are disabled.